backend/requirements.txt
# Stock API Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pandas>=2.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
finance-datareader>=0.9.50          # 한국/미국 주식 데이터 수집
pykrx>=1.0.0                       # KRX 종목 리스트 폴백용
yfinance>=0.2.50                    # 시장 지표 수집 (지수/원자재/환율/암호화폐)
requests>=2.31.0                    # HTTP 클라이언트 (수집 스크립트용)
pandas-datareader>=0.10.0          # FRED API (미국 거시경제 지표)

# === Phase 2: DB 연결 ===
asyncpg>=0.29.0                    # TimescaleDB(PostgreSQL) 비동기 드라이버
psycopg2-binary>=2.9.0             # 수집 스크립트용 동기 드라이버
python-dotenv>=1.0.0               # .env 환경변수 로딩

# === Phase 3: 인증 ===
python-jose[cryptography]>=3.3.0   # JWT 토큰 생성/검증
httpx>=0.27.0                      # Google OAuth API 호출

# === Phase 4: 백테스팅 ===
celery[rabbitmq]>=5.3.0            # 비동기 작업 큐
redis>=5.0.0                       # 결과 저장소
sqlalchemy>=2.0.0                  # pd.read_sql용 DB 연결 (tasks/backtest.py)